+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# wrapr 2.0.7 2021-02-03
+
+ * Add bc() blank-separated parsing, suggested by Emil Erik Pula Bellamy Begtrup-Bright \url{https://github.com/WinVector/wrapr/issues/12}.
+ * Fix strsplit_capture empty-return type.
+
 # wrapr 2.0.6 2020-12-06
 
  * Bump version.